requirements.txt
wget~=3.2
requests~=2.24.0
distro~=1.5.0
delegator.py~=0.1.1
python-crontab~=2.5.1
configobj~=5.0.6